Move visit() export from 'yaml/util' to 'yaml'
