+++ src/index.js
@@ -4,6 +4,7 @@ import { Document } from './doc/Document.js'
 import { YAMLSemanticError } from './errors.js'
 import { warn } from './log.js'
 
+export { visit } from './ast/index.js'
 export { defaultOptions, scalarOptions } from './options.js'
 export { Document, parseCST }
