+++ util.js
@@ -2,7 +2,6 @@ const util = require('./dist/util')
 
 exports.findPair = util.findPair
 exports.toJSON = util.toJSON
-exports.visit = util.visit
 
 exports.stringifyNumber = util.stringifyNumber
 exports.stringifyString = util.stringifyString
