+++ util.mjs
@@ -2,7 +2,6 @@ import util from './dist/util.js'
 
 export const findPair = util.findPair
 export const toJSON = util.toJSON
-export const visit = util.visit
 
 export const stringifyNumber = util.stringifyNumber
 export const stringifyString = util.stringifyString
